requirements.txt
Django==5.1
requests==2.32.3
numpy==2.1.0
djangorestframework==3.15.2
django-leaflet==0.30.1
